+++ CHANGELOG.md
@@ -9,6 +9,7 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 ### Fixed
 
 - [#286](https://github.com/llaville/php-compat-info/issues/286) - Mass method signature mismatch
+- [#292](https://github.com/llaville/php-compat-info/issues/292) - Uncaught Error: Object of class PhpParser\Node\UnionType could not be converted to string
 
 ## [5.4.2] - 2020-11-20